Refactor photo page Co-authored-by: Duda Melo <dudamelo82@gmail.com>"
