+++ app/src/pages/authPages/Photo/index.js
@@ -1,27 +1,13 @@
-import React, { useState } from 'react';
-import {
-    Image,
-    ImageBackground,
-    Text,
-    TouchableOpacity,
-    View,
-} from 'react-native';
+import React from 'react';
+import { ImageBackground, Text, TouchableOpacity, View } from 'react-native';
 import * as ImagePicker from 'expo-image-picker';
-import { Icon, CheckBox } from 'react-native-elements';
+import { Icon } from 'react-native-elements';
 import styles from './styles';
 import Container from '../../../components/Container';
-import TermsModal from '../../../components/modals/conditionTermsModal';
-import PrivacyPolicyModal from '../../../components/modals/privacyPolicyModal';
-import Buttom from '../../../components/UI/button';
+
 import { alertMessage } from '../../../utils/Alert';
 export default function Photo({ route, navigation }) {
     const { userDataFromAddressPage } = route.params;
-
-    const [selectedImage, setSelectedImage] = useState(null);
-    const [photo, setPhoto] = useState('');
-    const [termsModalVisible, setTermsModalVisible] = useState(false);
-    const [checked, setChecked] = useState(false);
-    const [privacyModalVisible, setPrivacyModalVisible] = useState(false);
 
     async function openImagePickerAsync() {
         const permissionResult = await ImagePicker.requestCameraRollPermissionsAsync();
@@ -40,122 +26,38 @@ export default function Photo({ route, navigation }) {
             return;
         }
 
-        setSelectedImage({
-            localUri: pickerResult.uri,
+        navigation.navigate('photoPreview', {
+            selectedPhoto: pickerResult.base64,
+            userDataFromAddressPage,
         });
-
-        setPhoto(pickerResult.base64);
     }
 
-    async function cancelHandler() {
-        setSelectedImage(null);
-        setPhoto('');
-    }
-    async function continueHandler() {
-        const userDataFromPhotoPage = {
-            photo,
-            ...userDataFromAddressPage,
-        };
-        navigation.navigate('riskGroup', { userDataFromPhotoPage });
-    }
-
-    const titleCheckBox = (
-        <View style={styles.checkBoxTitle}>
-            <View style={styles.checkBoxContent}>
-                <Text style={styles.checkBoxText}> Li e concordo com os </Text>
-                <TouchableOpacity onPress={() => setTermsModalVisible(true)}>
-                    <Text style={styles.hyperLink}> Termos de Uso </Text>
-                </TouchableOpacity>
-            </View>
-            <View style={styles.checkBoxContent}>
-                <Text style={styles.checkBoxText}>e as </Text>
-                <TouchableOpacity onPress={() => setPrivacyModalVisible(true)}>
-                    <Text style={styles.hyperLink}>
-                        Políticas de privacidade
-                    </Text>
-                </TouchableOpacity>
-                <Text style={styles.checkBoxText}>.</Text>
-            </View>
-        </View>
-    );
     return (
         <View style={styles.container}>
-            {selectedImage === null ? (
-                <ImageBackground
-                    source={require('../../../images/catPhoto.png')}
-                    style={styles.logo}>
-                    <Container>
-                        <View style={styles.backIcon}>
-                            <TouchableOpacity
-                                onPress={() => navigation.goBack()}>
-                                <Icon name={'arrow-back'} color={'black'} />
-                            </TouchableOpacity>
-                        </View>
-                        <View style={styles.textView}>
-                            <Text style={styles.text}>
-                                Também precisamos de uma foto sua, é só clicar
-                                na camêra aqui em baixo!
-                            </Text>
-                        </View>
-                        <View style={styles.btnView}>
-                            <TouchableOpacity
-                                onPress={openImagePickerAsync}
-                                style={styles.button}>
-                                <Icon name={'camera-alt'} color="gray" />
-                            </TouchableOpacity>
-                        </View>
-                    </Container>
-                </ImageBackground>
-            ) : (
-                <View style={styles.container}>
-                    <Image
-                        source={{ uri: selectedImage.localUri }}
-                        style={styles.thumbnail}
-                    />
-                    <View style={styles.selectText}>
+            <ImageBackground
+                source={require('../../../images/catPhoto.png')}
+                style={styles.logo}>
+                <Container>
+                    <View style={styles.backIcon}>
+                        <TouchableOpacity onPress={() => navigation.goBack()}>
+                            <Icon name={'arrow-back'} color={'black'} />
+                        </TouchableOpacity>
+                    </View>
+                    <View style={styles.textView}>
                         <Text style={styles.text}>
-                            Clique em continuar para prosseguir com o cadastro,
-                            ou voltar para escolher outra foto.
+                            Também precisamos de uma foto sua, é só clicar na
+                            camêra aqui em baixo!
                         </Text>
                     </View>
-                    <View style={styles.checkBox}>
-                        <CheckBox
-                            title={titleCheckBox}
-                            style={styles.checkbox}
-                            iconRight
-                            size={28}
-                            checked={checked}
-                            onPress={() => setChecked(!checked)}
-                            onIconPress={() => setChecked(!checked)}
-                        />
+                    <View style={styles.btnView}>
+                        <TouchableOpacity
+                            onPress={openImagePickerAsync}
+                            style={styles.button}>
+                            <Icon name={'camera-alt'} color="gray" />
+                        </TouchableOpacity>
                     </View>
-
-                    <View style={styles.buttonPreview}>
-                        <Buttom
-                            title="Voltar"
-                            type="notSelected"
-                            press={() => {
-                                cancelHandler();
-                            }}
-                        />
-                        <Buttom
-                            disabled={!checked}
-                            title="Continuar"
-                            press={() => {
-                                continueHandler();
-                            }}
-                        />
-                    </View>
-                    <TermsModal
-                        visible={termsModalVisible}
-                        setVisible={setTermsModalVisible}
-                    />
-                    <PrivacyPolicyModal
-                        visible={privacyModalVisible}
-                        setVisible={setPrivacyModalVisible}
-                    />
-                </View>
-            )}
+                </Container>
+            </ImageBackground>
         </View>
     );
 }
