+++ app/src/pages/authPages/Photo/styles.js
@@ -20,32 +20,13 @@ const styles = StyleSheet.create({
         borderRadius: 100,
         alignItems: 'center',
     },
-    thumbnail: {
-        width: 200,
-        height: 200,
-        borderRadius: 125,
-        marginTop: 20,
-    },
+
     text: {
         ...fonts.subtitle,
         fontFamily: 'montserrat-semibold',
         textAlign: 'center',
     },
-    smallText: {
-        ...fonts.subtitle,
-        fontSize: 12,
-        fontFamily: 'montserrat-semibold',
-        textAlign: 'center',
-        marginTop: 5,
-        marginBottom: 5,
-    },
-    hyperLink: {
-        ...fonts.subtitle,
-        fontSize: 13,
-        fontFamily: 'montserrat-semibold',
-        textAlign: 'left',
-        color: colors.primary,
-    },
+
     textView: {
         flex: 6,
         margin: 16,
@@ -56,74 +37,10 @@ const styles = StyleSheet.create({
         alignItems: 'center',
         marginVertical: 40,
     },
-    buttonPreview: {
-        width: '100%',
-        flex: 1,
-        flexDirection: 'row',
-        alignItems: 'flex-end',
-        alignContent: 'space-between',
-        justifyContent: 'space-evenly',
-        marginTop: 30,
-    },
-    btn: {
-        marginHorizontal: 10,
-        borderWidth: 2,
-        borderColor: colors.primary,
-        borderRadius: 4,
-        width: 150,
-        alignItems: 'center',
-    },
-    btnText: {
-        fontSize: 20,
-        color: colors.primary,
-        padding: 10,
-        fontFamily: 'montserrat-medium',
-    },
-    btn1: {
-        marginHorizontal: 10,
-        borderWidth: 2,
-        borderColor: colors.primary,
-        backgroundColor: colors.primary,
-        borderRadius: 4,
-        width: 150,
-        alignItems: 'center',
-    },
-    btnText1: {
-        fontSize: 20,
-        color: colors.light,
-        padding: 10,
-        fontFamily: 'montserrat-medium',
-    },
-    selectText: {
-        flex: 3,
-        alignItems: 'center',
-        justifyContent: 'center',
-        marginHorizontal: 20,
-        marginVertical: 40,
-        width: '75%',
-    },
+
     backIcon: {
         alignItems: 'flex-start',
         marginTop: 5,
-    },
-    checkBox: {
-        justifyContent: 'center',
-        alignContent: 'space-between',
-    },
-    checkBoxTitle: {
-        flexDirection: 'column',
-        alignItems: 'center',
-        justifyContent: 'center',
-        marginRight: 20,
-    },
-    checkBoxContent: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        justifyContent: 'center',
-    },
-    checkBoxText: {
-        fontSize: 13,
-        fontFamily: 'montserrat-medium',
     },
 });
